Move backend application into server folder Move frontend application into client folder Update data.ts - refactor constants format - add more application data Update Dockerfile run command Update Dockerfile run command Update server.ts - add api prefix and api version into business logic Update users test

+++ server/tests/integration/users/users.test.ts
@@ -1,6 +1,6 @@
 describe("Test users api",  function() {
-    const API_URL = "http://localhost:4000/api/users";
-    const API_LOGIN = "http://localhost:4000/api/login";
+    const API_URL = "http://localhost:4000/api/v1/users";
+    const API_LOGIN = "http://localhost:4000/api/v1/login";
     test("Create user", async function() {
         const notValidUser = {
             email: "2",
@@ -16,13 +16,13 @@ describe("Test users api",  function() {
         })
         expect(notValidResponse.status).toBe(400);
         const user = {
-            email: "test_create_" + Math.random().toString(16).substr(2, 8) + "@abv.bg",
+            email: "test_create_user@abv.bg",
             password: "123456",
             role: "admin"
         }
         const response = await fetch(API_URL, {
             method: "POST",
-            body: JSON.stringify(user),  
+            body: JSON.stringify(user),
             headers: {
                 "Content-Type": "application/json"
             }
@@ -32,7 +32,7 @@ describe("Test users api",  function() {
             expect(response.status).toBe(201);
             json.then(async result => {
                 expect(result.message).toEqual("Successfully registration");
-                expect(result.data._id).not.toBeNull();
+                expect(result.status).toEqual("success");
             })
         }
     });
@@ -54,7 +54,7 @@ describe("Test users api",  function() {
             expect(response.status).toBe(201);
             json.then(async result => {
                 expect(result.message).toEqual("Successfully registration");
-                expect(result.data._id).not.toBeNull();
+                expect(result.status).toEqual("success");
                 const getUserResponse = await fetch(API_URL + "/" + result.data._id)
                 const userJson = await getUserResponse.json()
                 if (userJson.ok) {
@@ -62,9 +62,9 @@ describe("Test users api",  function() {
                         expect(user.role).toEqual("admin");
                         expect(user).toHaveProperty("_id");
                         expect(user).toHaveProperty("password");
-                        expect(user).toHaveProperty("email");    
+                        expect(user).toHaveProperty("email");
                     })
-                }    
+                }
             })
         }
     });
@@ -79,49 +79,6 @@ describe("Test users api",  function() {
         }
     });
     test("Update user", async function() {
-        const userWithoutAdminRole = {
-            email: "update_not_admin_" + Math.random().toString(16).substr(2, 8) + "@abv.bg",
-            password: "123456",
-            role: "user"
-        }
-        const responseUserWithoutAdminRole = await fetch(API_URL, {
-            method: "POST",
-            body: JSON.stringify(userWithoutAdminRole),
-            headers: {
-                "Content-Type": "application/json"
-            }
-        })
-        if (responseUserWithoutAdminRole.ok) {
-            const jsonUserWithoutAdminRole = responseUserWithoutAdminRole.json()
-            expect(responseUserWithoutAdminRole.status).toBe(201);
-            jsonUserWithoutAdminRole.then(async result => {                
-                expect(result.message).toEqual("Successfully registration");
-                expect(result.data._id).not.toBeNull();
-                const responseLogin = await fetch(API_LOGIN, {
-                    method: "POST",
-                    body: JSON.stringify(userWithoutAdminRole),
-                    headers: {
-                        "Content-Type": "application/json"
-                    }
-                })
-                const jsonLogin = responseLogin.json()
-                expect(responseLogin.status).toBe(200);
-                jsonLogin.then(async login => {
-                    expect(login.data.token).not.toBeNull();
-                    const responseUpdate = await fetch(API_URL + '/' + result.data._id, {
-                        method: 'PATCH',
-                        body: JSON.stringify(userWithoutAdminRole),
-                        headers: {
-                            "Content-Type": "application/json",
-                            "x-access-token": login.data.token
-                        }
-
-                    })
-                    expect(responseUpdate.status).toBe(401);
-                })
-            })
-            
-        }
         const userWithAdminRole = {
             email: "update_admin_" + Math.random().toString(16).substr(2, 8) + "@abv.bg",
             password: "123456",
@@ -138,7 +95,7 @@ describe("Test users api",  function() {
         const jsonUserWithAdminRole = responseUserWithAdminRole.json()
         jsonUserWithAdminRole.then(async result => {
             expect(result.message).toEqual("Successfully registration");
-            expect(result.data._id).not.toBeNull();
+            expect(result.status).toEqual("success");
             const responseLogin = await fetch(API_LOGIN, {
                 method: "POST",
                 body: JSON.stringify(userWithAdminRole),
@@ -150,14 +107,56 @@ describe("Test users api",  function() {
             const jsonLogin = responseLogin.json()
             jsonLogin.then(async login => {
                 expect(login.data.token).not.toBeNull();
-                const responseUpdate = await fetch(API_URL + '/' + result.data._id, {
+                const userWithoutAdminRole = {
+                    email: "update_not_admin_" + Math.random().toString(16).substr(2, 8) + "@abv.bg",
+                    password: "123456",
+                    role: "user"
+                }
+                const responseUserWithoutAdminRole = await fetch(API_URL, {
+                    method: "POST",
+                    body: JSON.stringify(userWithoutAdminRole),
+                    headers: {
+                        "Content-Type": "application/json"
+                    }
+                })
+                if (responseUserWithoutAdminRole.ok) {
+                    const jsonUserWithoutAdminRole = responseUserWithoutAdminRole.json()
+                    expect(responseUserWithoutAdminRole.status).toBe(201);
+                    jsonUserWithoutAdminRole.then(async result => {
+                        expect(result.message).toEqual("Successfully registration");
+                        expect(result.status).toEqual("success");
+                        const responseLogin = await fetch(API_LOGIN, {
+                            method: "POST",
+                            body: JSON.stringify(userWithoutAdminRole),
+                            headers: {
+                                "Content-Type": "application/json"
+                            }
+                        })
+                        const jsonLogin = responseLogin.json()
+                        expect(responseLogin.status).toBe(200);
+                        jsonLogin.then(async login => {
+                            expect(login.data.token).not.toBeNull();
+                            const responseUpdate = await fetch(API_URL + '/' + login.data.logged_user_id, {
+                                method: 'PATCH',
+                                body: JSON.stringify(userWithoutAdminRole),
+                                headers: {
+                                    "Content-Type": "application/json",
+                                    "x-access-token": login.data.token
+                                }
+
+                            })
+                            expect(responseUpdate.status).toBe(401);
+                        })
+                    })
+
+                }
+                const responseUpdate = await fetch(API_URL + '/' + login.data.logged_user_id, {
                     method: 'PATCH',
                     body: JSON.stringify(userWithAdminRole),
                     headers: {
                         "Content-Type": "application/json",
                         "x-access-token": login.data.token
                     }
-
                 })
                 expect(responseUpdate.status).toBe(200);
             })
@@ -165,43 +164,31 @@ describe("Test users api",  function() {
     });
     test("Delete user", async function() {
         const userWithAdminRole = {
-            email: "test_delete_user@abv.bg",
+            email: "test_admin_user@abv.bg",
             password: "123456",
             role: "admin"
         }
-        const responseUserWithAdminRole = await fetch(API_URL, {
+        const responseLogin = await fetch(API_LOGIN, {
             method: "POST",
             body: JSON.stringify(userWithAdminRole),
             headers: {
                 "Content-Type": "application/json"
             }
         })
-        expect(responseUserWithAdminRole.status).toBe(201);
-        const jsonUserWithAdminRole = responseUserWithAdminRole.json()
-        jsonUserWithAdminRole.then(async result => {
-            expect(result.message).toEqual("Successfully registration");
-            expect(result.data._id).not.toBeNull();
-            const responseLogin = await fetch(API_LOGIN, {
-                method: "POST",
-                body: JSON.stringify(userWithAdminRole),
-                headers: {
-                    "Content-Type": "application/json"
-                }
-            })
+        if (responseLogin.ok) {
             expect(responseLogin.status).toBe(200);
             const jsonLogin = responseLogin.json()
             jsonLogin.then(async login => {
                 expect(login.data.token).not.toBeNull();
-                const responseDelete = await fetch(API_URL + '/' + result.data._id, {
+                const responseDelete = await fetch(API_URL + '/' + login.data.logged_user_id, {
                     method: 'DELETE',
                     headers: {
                         "Content-Type": "application/json",
                         "x-access-token": login.data.token
                     }
-
                 })
-                expect(responseDelete.status).toBe(200);
+                expect(responseDelete.status).toBe(204);
             })
-        });
-    });
+        }
+    })
 });
